+++ scripts/create_market_omen.py
@@ -3,10 +3,12 @@ from datetime import datetime
 import typer
 
 from prediction_market_agent_tooling.gtypes import PrivateKey, xdai_type
-from prediction_market_agent_tooling.markets.omen.omen import (
-    OMEN_DEFAULT_MARKET_FEE,
+from prediction_market_agent_tooling.markets.omen.data_models import (
     OMEN_FALSE_OUTCOME,
     OMEN_TRUE_OUTCOME,
+)
+from prediction_market_agent_tooling.markets.omen.omen import (
+    OMEN_DEFAULT_MARKET_FEE,
     omen_create_market_tx,
 )
 from prediction_market_agent_tooling.tools.web3_utils import verify_address
